% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change_from_target_exp_1.R
\name{bar_plot_change_from_target}
\alias{bar_plot_change_from_target}
\title{Create and save a bar plot of percentage change in feature representation from target.}
\usage{
bar_plot_change_from_target(aa, parms)
}
\arguments{
\item{aa}{the dataframe with the data to plot.}

\item{parms}{a list of parameters.}
}
\description{
Create and save a bar plot of percentage change in feature representation from target.
}
