% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_representation_error_exp_1.R
\name{bar_plot_representation_error}
\alias{bar_plot_representation_error}
\title{Create and save a bar plot of feature representation error.}
\usage{
bar_plot_representation_error(aa, parms)
}
\arguments{
\item{aa}{the dataframe with the data to plot.}

\item{parms}{a list of parameters.}
}
\description{
Create and save a bar plot of feature representation error.
}
