% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_proportions_exp_1.R
\name{bar_plot_species_proportions}
\alias{bar_plot_species_proportions}
\title{Create and save a bar plot of the proportion of species above and below target.}
\usage{
bar_plot_species_proportions(aa, parms)
}
\arguments{
\item{aa}{the dataframe with the data to plot.}

\item{parms}{a list of parameters.}
}
\description{
Create and save a bar plot of the proportion of species above and below target.
}
