% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{cal_norm_upper_bound}
\alias{cal_norm_upper_bound}
\title{Normalize experts' upper bounds.}
\usage{
cal_norm_upper_bound(best_g, upper_b, confid, possib_lev)
}
\arguments{
\item{best_g}{the expert's best guess.}

\item{upper_b}{the expert's upper bound.}

\item{confid}{the expert's confidence level.}

\item{possib_lev}{the possibility level to normalize to.}
}
\description{
It applies the normalization formula described in Mcbride et al. 2012.
}
