% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{calc_daly_prop_selected}
\alias{calc_daly_prop_selected}
\title{Calculate the proportion of the study region which has been selected.}
\usage{
calc_daly_prop_selected(sel_pu_ind, area_file)
}
\arguments{
\item{sel_pu_ind}{the output of \code{get_selected_pu_index()}.}

\item{area_file}{the matrix of treated area for each threat (columns) at each site (rows).}
}
\description{
Calculate the proportion of the study region which has been selected.
}
