% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_uncertainty_analysis.R
\name{calc_uncertainty_diagnostics}
\alias{calc_uncertainty_diagnostics}
\title{Average uncertainty diagnostics across species.}
\usage{
calc_uncertainty_diagnostics(x)
}
\arguments{
\item{the}{output of \code{calculate_representation_error()}}
}
\description{
Average uncertainty diagnostics across species.
}
