% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_uncertainty_analysis.R
\name{calculate_representation_error}
\alias{calculate_representation_error}
\title{Calculate a set of diagnostics for a specific assumption regarding reponses
  uncertanity.}
\usage{
calculate_representation_error(x, b, c, parms, cons_features,
  site_threat_array_cat, all_run_results, action_cost_list,
  responses_to_actions_EXP, required_actions, site_species_array,
  output_by_species)
}
\arguments{
\item{b}{a list of species-action benefits calculated assuming that the true
species responses corresponds to the experts' best guesses, the lower
bounds and the upper bounds.}

\item{c}{a character vector of name tags.}

\item{site_threat_array_cat}{a matrix of the category of magnitute for each threat
in each site.}

\item{all_run_results}{a list of solution objects.}

\item{responses_to_actions_EXP}{a list of matrices with the benefit of selecting each
level of effort. Each matrix stores the benefit of implementing each level of effort
(rows) for different threat category at the site (columns) for each site-action
combination (list slot). The list has lenght equals to the combination of site
and actions. Response values are calculated assuming that the true species
responses corresponds to the experts' best guesses.}

\item{required_actions}{a list of the total number of threats (rows) affecting
each species (columns) at each site (list slot).}

\item{site_species_array}{a matrix of the area of occupancy of each feature (columns)
in each site (rows).}

\item{output_by_species}{a logical indicating wheather a table for each species is saved.}

\item{a}{solution object. A list.}
}
\description{
Calculate a set of diagnostics for a specific assumption regarding reponses
  uncertanity.
}
