% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_OF.R
\name{components_OF}
\alias{components_OF}
\title{Calculate the value of the objective function.}
\usage{
components_OF(parameters, cons_feat_array, all_site_action_int_combs,
  site_action_array, action_costs, site_threat_array_cat, responses_to_actions,
  site_species_array, required_actions)
}
\arguments{
\item{parameters}{a list of parameters.}

\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}

\item{all_site_action_int_combs}{a matrix of all the combinations (rows) of site,
action and level of effort (columns) available for selection in the study area.}

\item{site_action_array}{a matrix of the selected effort for each action (columns)
in each site (rows).}

\item{action_costs}{a list of matrices with the cost of selecting each level of effort.
Each matrix stores the cost of implementing each level of effort (rows) for each
action (columns) in each site (list slot). The list has lenght equals to the
number of site.}

\item{site_threat_array_cat}{a matrix of the category of magnitute for each threat
in each site.}

\item{responses_to_actions}{a list of matrices with the benefit of selecting each
level of effort. Each matrix stores the benefit of implementing each level of effort
(rows) for different threat category at the site (columns) for each site-action
combination (list slot). The list has lenght equals to the combination of site
and actions.}

\item{site_species_array}{a matrix of the area of occupancy of each feature (columns)
in each site (rows).}

\item{required_actions}{a list of the total number of threats (rows) affecting
each species (columns) at each site (list slot).}
}
\description{
Calculate the value of the objective function.
}
