% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_persistence.R
\name{count.persistence}
\alias{count.persistence}
\title{Calculate the benefit of selected effort for each species.}
\usage{
count.persistence(site, action, site_action_array, action_costs,
  site_threat_array_cat, responses_to_actions)
}
\arguments{
\item{site}{the numeric id of the selected site.}

\item{action}{the numeric id of the selected action.}

\item{site_action_array}{a matrix of the selected effort for each action (columns)
in each site (rows).}

\item{action_costs}{a list of matrices with the cost of selecting each level of effort.
Each matrix stores the cost of implementing each level of effort (rows) for each
action (columns) in each site (list slot). The list has lenght equals to the
number of site.}

\item{site_threat_array_cat}{a matrix of the category of magnitute for each threat
in each site.}

\item{responses_to_actions}{a list of matrices with the benefit of selecting each
level of effort. Each matrix stores the benefit of implementing each level of effort
(rows) for different threat category at the site (columns) for each site-action
combination (list slot). The list has lenght equals to the combination of site
and actions.}
}
\description{
Calculate the benefit of selected effort for each species.
}
