% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_exp_design.R
\name{create_exp_des}
\alias{create_exp_des}
\title{Create a set of parameter combinations to run the prioritization on.}
\usage{
create_exp_des(parms)
}
\arguments{
\item{parms}{a list of parameters.}
}
\description{
Create a set of parameter combinations to run the prioritization on.
}
