% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{get.action.costs}
\alias{get.action.costs}
\title{Create action cost list.}
\usage{
get.action.costs(site_threat_array_cat, planning_unit = NULL, no.levels = 3)
}
\arguments{
\item{site_threat_array_cat}{a matrix of the category of magnitute for each threat
in each site.}

\item{planning_unit}{a matrix of the cost of abating completely each threat (columns)
in each site.}

\item{no.levels}{Number of threat categories (3).}
}
\description{
Create a list with the costs of all levels of effort available for each action at each site.
}
