% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{get.required_actions}
\alias{get.required_actions}
\title{Create required action list}
\usage{
get.required_actions(site_threat_array_cat, responses_to_actions,
  cons_feat_array)
}
\arguments{
\item{site_threat_array_cat}{a matrix of the category of magnitute for each threat
in each site.}

\item{responses_to_actions}{a list of matrices with the benefit of selecting each
level of effort. Each matrix stores the benefit of implementing each level of effort
(rows) for different threat category at the site (columns) for each site-action
combination (list slot). The list has lenght equals to the combination of site
and actions.}

\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}
}
\description{
It creates a list of the actions required to abate all threats to each species.
}
