% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{get.site.action.intensities.combs}
\alias{get.site.action.intensities.combs}
\title{Create site action intensity combination array.}
\usage{
get.site.action.intensities.combs(parms, site_threat_array_cat, no.levels = 3)
}
\arguments{
\item{site_threat_array_cat}{a matrix of the category of magnitute for each threat
in each site.}

\item{no.levels}{Number of threat categories (3).}
}
\description{
Create site action intensity combination array.
}
