% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{get.threat.intensity.category}
\alias{get.threat.intensity.category}
\title{Create threat intensity category array.}
\usage{
get.threat.intensity.category(parms, site_threat_array)
}
\arguments{
\item{parms}{a list of parameters.}

\item{site_threat_array}{a matrix of the magnitute of each threat in each site.}
}
\description{
Create a matrix of the category of magnitute (0, 1, 2 or 3) for each threat in each site.
}
