% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{get_cost_breakdown}
\alias{get_cost_breakdown}
\title{Create table 2 for paper.}
\usage{
get_cost_breakdown(site_action_array, treated_area_file, action_cost_list,
  scaling_factors)
}
\arguments{
\item{site_action_array}{a matrix of the selected effort for each action (columns)
in each site (rows).}

\item{scaling_factors}{a list of action intensity scaling factors.}
}
\description{
Create table 2 for paper.
}
