% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{get_run_diagnostics}
\alias{get_run_diagnostics}
\title{Get diagnostics from a prioritization solution.}
\usage{
get_run_diagnostics(solution)
}
\arguments{
\item{solution}{a solution object. A list.}
}
\description{
Get diagnostics from a prioritization solution.
}
