% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{get_selected_pu_index}
\alias{get_selected_pu_index}
\title{Get indices of selected planning units.}
\usage{
get_selected_pu_index(solution)
}
\arguments{
\item{solution}{a solution object. A list.}
}
\description{
The function returns the indices of the planning units which have been selected at the end of the annealing.
}
