% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{moving_index_search}
\alias{moving_index_search}
\title{Get indices of best runs}
\usage{
moving_index_search(my_data, runs)
}
\arguments{
\item{my_data}{a data frame with summary diagnotics for each experiment run.}

\item{runs}{number of replicates in the experiment.}
}
\description{
The function returns the indices of the runs with lowest objective function value among their replicate runs.
}
