% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_run.R
\name{one_run}
\alias{one_run}
\title{Reshape the data in the desired form and feed them to the function which runs
  the prioritization. Collects the outputs as well.}
\usage{
one_run(x, parms, site_threat_array, planning_unit, cons_feat_array,
  site_species_array, species_responses)
}
\arguments{
\item{x}{a one line data frame with the parameters to be used in the run.}

\item{parms}{a list of parameters.}

\item{planning_unit}{a matrix of the cost of abating completely each threat (columns)
in each site.}

\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}

\item{site_species_array}{a matrix of the area of occupancy of each feature (columns)
in each site (rows).}
}
\description{
Reshape the data in the desired form and feed them to the function which runs
  the prioritization. Collects the outputs as well.
}
