% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change_from_target_exp_2.R
\name{plot_change_from_target}
\alias{plot_change_from_target}
\title{Create and save a scatter plot of percentage change in feature representation from target,
  for different target levels.}
\usage{
plot_change_from_target(aa, parms)
}
\arguments{
\item{aa}{the dataframe with the data to plot.}

\item{parms}{a list of parameters.}
}
\description{
Create and save a scatter plot of percentage change in feature representation from target,
  for different target levels.
}
