% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effort_map.R
\name{plot_effort_map}
\alias{plot_effort_map}
\title{Create and save a map of selected effort for each action.}
\usage{
plot_effort_map(solution, catch_shp, river_shp, catch_shp_outline)
}
\arguments{
\item{solution}{a solution object. A list.}

\item{catch_shp}{the shape file of the subcatchments.}

\item{river_shp}{the shape file of the rivers.}

\item{catch_shp_outline}{the shape file of the study area border.}
}
\description{
Create and save a map of selected effort for each action.
}
