% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_representation_error_exp_2.R
\name{plot_representation_error}
\alias{plot_representation_error}
\title{Create and save a scatter plot of feature representation error
  for different target levels.}
\usage{
plot_representation_error(aa, parms)
}
\arguments{
\item{aa}{the dataframe with the data to plot.}

\item{parms}{a list of parameters.}
}
\description{
Create and save a scatter plot of feature representation error
  for different target levels.
}
