% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_proportions_exp_2.R
\name{plot_species_proportions}
\alias{plot_species_proportions}
\title{Create and save a scatter plot of the proportion of species above and below
 target for different target levels.}
\usage{
plot_species_proportions(aa, parms)
}
\arguments{
\item{aa}{the dataframe with the data to plot.}

\item{parms}{a list of parameters.}
}
\description{
Create and save a scatter plot of the proportion of species above and below
 target for different target levels.
}
