% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trackers.R
\name{plot_trackers}
\alias{plot_trackers}
\title{Plot temperature, cost and feature penalty during the annealing routine.}
\usage{
plot_trackers(solution)
}
\arguments{
\item{solution}{a solution object. A list.}
}
\description{
Plot temperature, cost and feature penalty during the annealing routine.
}
