% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{set_fixed_targets}
\alias{set_fixed_targets}
\title{Set fixed feature targets.}
\usage{
set_fixed_targets(cons_feat_array, site_species_array, target_level)
}
\arguments{
\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}

\item{site_species_array}{a matrix of the area of occupancy of each feature (columns)
in each site (rows).}

\item{target_level}{The target level. Numeric.}
}
\description{
Calculate constant feature targets for all features. It modifies the \code{cons_feat_array}.
}
