% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_functions.R
\name{set_scaled_targets}
\alias{set_scaled_targets}
\title{Set area-scaled targets.}
\usage{
set_scaled_targets(cons_feat_array, occurrence_limits, target_limits)
}
\arguments{
\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}

\item{occurrence_limits}{a numeric vector of the occurrence limits of the interpolation.}

\item{target_limits}{a numeric vector of the target limits of the interpolation.}
}
\description{
Calculate feature targets which are linearly scaled to each feature's area of occupancy.
  It uses linear interpolation and the function \code{approx()}. It modifies the
  \code{cons_feat_array}.
}
