% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_run.R
\name{wrapper}
\alias{wrapper}
\title{Wrapper for vectorizing \code{one_run()}.}
\usage{
wrapper(parms, exp_des, site_threat_array, planning_unit, cons_feat_array,
  site_species_array, species_responses, parallel)
}
\arguments{
\item{parms}{a list of parameters.}

\item{exp_des}{the fulls set of parameter combinations to run the prioritization on.}

\item{planning_unit}{a matrix of the cost of abating completely each threat (columns)
in each site.}

\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}

\item{site_species_array}{a matrix of the area of occupancy of each feature (columns)
in each site (rows).}

\item{parallel}{a logical determining wheather the vectorization is performed in parallel.}
}
\description{
Wrapper for vectorizing \code{one_run()}.
}
