% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_uncertainty_analysis.R
\name{wrapper_to_get_responses}
\alias{wrapper_to_get_responses}
\title{Wrapper for vectorizing \code{get.responses.to.actions()}.}
\usage{
wrapper_to_get_responses(i, species_responses, cons_feat_array)
}
\arguments{
\item{i}{the numeric index of the vector}

\item{cons_feat_array}{a matrix of information (columns) for each conservation
feature (rows). Information include: area of occupancy, target level, species
penalty factor (spf), faunal group id and ecological group id.}
}
\description{
Wrapper for vectorizing \code{get.responses.to.actions()}.
}
